refactor: DefaultRoleButton, RoleInfoRow, SearchNavBar에 메모이제이션 추가
