+++ src/client/src/setting/server/role/role-information/components/SearchNavBar.jsx
@@ -61,4 +61,4 @@ function SearchNavBar() {
   );
 }
 
-export default SearchNavBar;
+export default React.memo(SearchNavBar);
